requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.27.0
openai==1.30.0
pytrends==4.9.2
gspread==6.0.2
google-auth==2.29.0
apscheduler==3.10.4
scikit-learn==1.4.2
nltk==3.8.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
lxml==5.2.2
